+++ dataloom/tests/sqlite3/test_connection_uri_sqlite.py
@@ -1,13 +1,16 @@
 import pytest
 
 
-class TestConnectionSQLite:
+class TestConnectionURISQLite:
     def test_connect_with_wrong_dialect(self):
         from dataloom import Loom
         from dataloom.exceptions import UnsupportedDialectException
+        from dataloom.keys import SQLiteConfig
 
         with pytest.raises(UnsupportedDialectException) as exc_info:
-            sqlite_loom = Loom(dialect="hay", database="hi.db")
+            sqlite_loom = Loom(
+                dialect="hay", connection_uri=SQLiteConfig.connection_uri
+            )
             conn = sqlite_loom.connect()
             conn.close()
 
@@ -18,8 +21,9 @@ class TestConnectionSQLite:
 
     def test_connect_correct_connection(self):
         from dataloom import Loom
+        from dataloom.keys import SQLiteConfig
 
-        sqlite_loom = Loom(dialect="sqlite", database="hi.db")
+        sqlite_loom = Loom(dialect="sqlite", connection_uri=SQLiteConfig.connection_uri)
         conn = sqlite_loom.connect()
         conn.close()
         assert conn is not None